Add .idea files and modify others for project setup
This commit includes IntelliJ-based configuration files, such as .idea/prettier.xml and .idea/vcs.xml, and the project register file .idea/modules.xml, and project specific files like .idea/izipizza.iml. The addition of .idea/.gitignore is to exclude files in a JetBrains project that should be kept local.

Also, it includes the creation of a file under 'src/app/linkdestinationmagnet' which contains a function to redirect the user to a specific Instagram reel, and the removal of several CSS rules and HTML elements from three files in 'src/app/'.

All these changes are made to improve project data, layout, and SEO metadata - a necessary step towards setting up the project for working efficiency and personalization.